cache_decorator: fixed @memoized decorator to use a global cache
Even though @memoized is supposed to cache the constructed class variables
when used applied to V1Meta.asset_class(), it doesn't, because
the data variable only contains a single instance, thus the cache
only lasts until the next differing asset type is instantiated. This
caused a lot of extra round-trips between the client and server in
complex processing.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>