v1meta: handle relations in returned data properly
When relations and attributes related to relations are returned by a
REST query, they should be used to construct the associated instances, so
that no extra round trips are necessary when these relations are then
used by the querying code.

For instance, this code:

for actual in  v1meta.Actual.select('Workitem.Parent.Name'):
	actual.Workitem.Parent.Name

caused an extra round trip at the time of the reference, since
actual.Workitem.Parent is a relation that wasn't properly parsed into
the "actual" asset.

With these changes, this is now done, so the loop above got a whole lot
faster.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>