V1Query.select: add containing relations automatically to the select set
In case a select contains an attribute of a relation, automatically include
a reference to the container in the select set.

For example this patch makes this work:

for actual in v1.Actual.select("Workitem.Parent.Name"):
	print actual.Workitem.Parent.Name

Without this patch, Workitem, Workitem.Parent would have to be added to
the argument list manually.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>