base_asset: fixed AsOf member
The historic date a given object refers to was added to an object using
with_data(), however as it is missing from the meta information returned
by V1, it is not registered as an attribute, thus unavailable through
the object.

The solution is to use a normal instance variable, which can be set to
None when we are not querying the history.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>