Updated client.py to use oauth when Granted access tokens are used to connect to the versioneone server. Modified methods: get_asset_xml; get_query_xml; execute_operation; get_attr; create_asset; and update_asset
Updated query.py to accept a token_flag which indicates that the user is
using Access token authentication as opposed to user/pw credential
management. Also modified run_single_query method to check the token
param in the V1Query object to determine if it should use oauth in the
URL path for the query.